requirements.txt
python-telegram-bot>=12.0
lyricfetch>=1.1.0
git+https://github.com/plamere/spotipy.git
flask
unidecode
